Add @tryghost/verification-trigger
refs: https://github.com/TryGhost/Toolbox/issues/166

Replace previous attempt at adding a verification trigger module with a package. Since it handles API interactions using events, it doesn't need to be injected into the API service.
