+++ core/server/services/members/api.js
@@ -13,7 +13,6 @@ const SingleUseTokenProvider = require('./SingleUseTokenProvider');
 const urlUtils = require('../../../shared/url-utils');
 const labsService = require('../../../shared/labs');
 const offersService = require('../offers');
-const VerificationTrigger = require('./verification/verification-trigger');
 
 const MAGIC_LINK_TOKEN_VALIDITY = 24 * 60 * 60 * 1000;
 
@@ -195,12 +194,7 @@ function createApiInstance(config) {
         },
         stripeAPIService: stripeService.api,
         offersAPI: offersService.api,
-        labsService: labsService,
-        verificationTrigger: new VerificationTrigger({
-            settingsCache,
-            urlUtils,
-            ghostMailer
-        })
+        labsService: labsService
     });
 
     return membersApiInstance;
